+++ frontend/src/metabase/visualizations/components/ClickActions/ClickActionsPopover.unit.spec.tsx
@@ -1,133 +1,23 @@
 import userEvent from "@testing-library/user-event";
-import { waitFor } from "@testing-library/react";
+import { getIcon, queryIcon, screen } from "__support__/ui";
 import {
-  getIcon,
-  queryIcon,
-  renderWithProviders,
-  screen,
-} from "__support__/ui";
-import {
-  createOrdersCreatedAtDatasetColumn,
-  createOrdersCreatedAtField,
-  createOrdersDiscountDatasetColumn,
-  createOrdersDiscountField,
-  createOrdersIdDatasetColumn,
-  createOrdersProductIdDatasetColumn,
-  createOrdersProductIdField,
-  createOrdersQuantityDatasetColumn,
-  createOrdersQuantityField,
-  createOrdersSubtotalDatasetColumn,
-  createOrdersSubtotalField,
-  createOrdersTable,
-  createOrdersTableDatasetColumns,
-  createOrdersTableDatasetColumnsMap,
-  createOrdersTaxDatasetColumn,
-  createOrdersTaxField,
-  createOrdersTotalDatasetColumn,
-  createOrdersTotalField,
-  createOrdersUserIdDatasetColumn,
-  createOrdersUserIdField,
-  createSampleDatabase,
   ORDERS,
   ORDERS_ID,
-  PEOPLE,
-  PEOPLE_ID,
-  PRODUCTS,
-  PRODUCTS_ID,
   SAMPLE_DB_ID,
 } from "metabase-types/api/mocks/presets";
-import { ClickActionsPopover } from "metabase/visualizations/components/ClickActions/ClickActionsPopover";
-import type { RegularClickAction } from "metabase/visualizations/types";
-import { getMode } from "metabase/visualizations/click-actions/lib/modes";
-import { checkNotNull } from "metabase/lib/types";
-import type {
-  DatasetColumn,
-  DatasetQuery,
-  Filter,
-  RowValue,
-  Series,
-  Card,
-} from "metabase-types/api";
+import type { DatasetQuery, Filter, RowValue } from "metabase-types/api";
 import registerVisualizations from "metabase/visualizations/register";
 import { POPOVER_TEST_ID } from "metabase/visualizations/click-actions/actions/ColumnFormattingAction/ColumnFormattingAction";
-import { createMockSingleSeries } from "metabase-types/api/mocks";
-import { ZOOM_IN_ROW } from "metabase/query_builder/actions";
-import { createMockMetadata } from "__support__/metadata";
-import type { ClickObject } from "metabase-lib/queries/drills/types";
 import { SAMPLE_METADATA } from "metabase-lib/test-helpers";
 import Question from "metabase-lib/Question";
-import type StructuredQuery from "metabase-lib/queries/StructuredQuery";
-import type Dimension from "metabase-lib/Dimension";
 import {
   ORDERS_COLUMNS,
-  ORDERS_COLUMNS_LIST,
   ORDERS_DATASET_QUERY,
   ORDERS_ROW_VALUES,
 } from "metabase-lib/tests/drills-common";
+import { setup } from "./tests/common";
 
 registerVisualizations();
-
-const ORDERS_METADATA_WITH_MULTIPLE_PK = createMockMetadata({
-  databases: [
-    createSampleDatabase({
-      tables: [
-        createOrdersTable({
-          fields: [
-            createOrdersUserIdField({
-              semantic_type: "type/PK",
-            }),
-            createOrdersProductIdField({
-              semantic_type: "type/PK",
-            }),
-            createOrdersSubtotalField(),
-            createOrdersTaxField(),
-            createOrdersTotalField(),
-            createOrdersDiscountField(),
-            createOrdersCreatedAtField(),
-            createOrdersQuantityField(),
-          ],
-        }),
-      ],
-    }),
-  ],
-});
-const ORDERS_QUESTION_WITH_MULTIPLE_PK = Question.create({
-  metadata: ORDERS_METADATA_WITH_MULTIPLE_PK,
-  dataset_query: {
-    type: "query",
-    database: SAMPLE_DB_ID,
-    query: {
-      "source-table": ORDERS_ID,
-    },
-  },
-});
-const ORDERS_COLUMNS_WITH_MULTIPLE_PK = {
-  USER_ID: createOrdersUserIdDatasetColumn({
-    semantic_type: "type/PK",
-  }),
-  PRODUCT_ID: createOrdersProductIdDatasetColumn({
-    semantic_type: "type/PK",
-  }),
-  SUBTOTAL: createOrdersSubtotalDatasetColumn(),
-  TAX: createOrdersTaxDatasetColumn(),
-  TOTAL: createOrdersTotalDatasetColumn(),
-  DISCOUNT: createOrdersDiscountDatasetColumn(),
-  CREATED_AT: createOrdersCreatedAtDatasetColumn(),
-  QUANTITY: createOrdersQuantityDatasetColumn(),
-};
-const ORDERS_ROW_VALUES_WITH_MULTIPLE_PK: Record<
-  keyof typeof ORDERS_COLUMNS_WITH_MULTIPLE_PK,
-  RowValue
-> = {
-  USER_ID: "1",
-  PRODUCT_ID: "105",
-  SUBTOTAL: 52.723521442619514,
-  TAX: 2.9,
-  TOTAL: 49.206842233769756,
-  DISCOUNT: null,
-  CREATED_AT: "2025-12-06T22:22:48.544+02:00",
-  QUANTITY: 2,
-};
 
 describe("ClickActionsPopover", function () {
   describe("apply click actions", () => {
@@ -466,356 +356,8 @@ describe("ClickActionsPopover", function () {
         },
       );
     });
-
-    describe("ObjectDetailsZoomDrill", () => {
-      it.each([
-        {
-          column: ORDERS_COLUMNS.ID,
-          columnName: ORDERS_COLUMNS.ID.name,
-          cellValue: ORDERS_ROW_VALUES.ID,
-          expectedAction: {
-            type: ZOOM_IN_ROW,
-            payload: { objectId: ORDERS_ROW_VALUES.ID },
-          },
-        },
-        {
-          column: ORDERS_COLUMNS.TOTAL,
-          columnName: ORDERS_COLUMNS.TOTAL.name,
-          cellValue: ORDERS_ROW_VALUES.TOTAL,
-          expectedAction: {
-            type: ZOOM_IN_ROW,
-            payload: { objectId: ORDERS_ROW_VALUES.ID },
-          },
-        },
-      ])(
-        "should apply drill on $columnName cell click",
-        async ({ column, cellValue, expectedAction }) => {
-          const { props } = await setup({
-            clicked: {
-              column,
-              value: cellValue,
-              data: ORDERS_COLUMNS_LIST.map(column => ({
-                col: ORDERS_COLUMNS[
-                  column.name as keyof typeof ORDERS_ROW_VALUES
-                  ],
-                value:
-                  ORDERS_ROW_VALUES[
-                    column.name as keyof typeof ORDERS_ROW_VALUES
-                    ],
-              })),
-            },
-          });
-
-          const drill = screen.getByText("View details");
-          expect(drill).toBeInTheDocument();
-
-          userEvent.click(drill);
-
-          expect(props.dispatch).toHaveBeenCalledTimes(3);
-          expect(props.dispatch).toHaveBeenCalledWith(expectedAction);
-        },
-      );
-    });
-
-    describe("ObjectDetailsFkDrill", () => {
-      it.each([
-        {
-          column: ORDERS_COLUMNS.USER_ID,
-          columnName: ORDERS_COLUMNS.USER_ID.name,
-          cellValue: ORDERS_ROW_VALUES.USER_ID,
-          expectedCard: {
-            dataset_query: {
-              database: SAMPLE_DB_ID,
-              query: {
-                filter: [
-                  "=",
-                  [
-                    "field",
-                    PEOPLE.ID,
-                    {
-                      "base-type": "type/BigInteger",
-                    },
-                  ],
-                  ORDERS_ROW_VALUES.USER_ID,
-                ],
-                "source-table": PEOPLE_ID,
-              },
-              type: "query",
-            },
-            display: "table",
-          },
-        },
-        {
-          column: createOrdersProductIdDatasetColumn(),
-          columnName: createOrdersProductIdDatasetColumn().name,
-          cellValue: ORDERS_ROW_VALUES.PRODUCT_ID,
-          expectedCard: {
-            dataset_query: {
-              database: SAMPLE_DB_ID,
-              query: {
-                filter: [
-                  "=",
-                  [
-                    "field",
-                    PRODUCTS.ID,
-                    {
-                      "base-type": "type/BigInteger",
-                    },
-                  ],
-                  ORDERS_ROW_VALUES.PRODUCT_ID,
-                ],
-                "source-table": PRODUCTS_ID,
-              },
-              type: "query",
-            },
-            display: "table",
-          },
-        },
-      ])(
-        "should apply drill on $columnName cell click",
-        async ({ column, cellValue, expectedCard }) => {
-          const { props } = await setup({
-            clicked: {
-              column,
-              value: cellValue,
-            },
-          });
-
-          const drill = screen.getByText("View details");
-          expect(drill).toBeInTheDocument();
-
-          userEvent.click(drill);
-
-          expect(props.onChangeCardAndRun).toHaveBeenCalledTimes(1);
-          expect(props.onChangeCardAndRun).toHaveBeenLastCalledWith(
-            expect.objectContaining({
-              nextCard: expect.objectContaining(expectedCard),
-            }),
-          );
-        },
-      );
-    });
-
-    describe("ObjectDetailsPkDrill", () => {
-      it.each<{
-        columnName: keyof typeof ORDERS_COLUMNS_WITH_MULTIPLE_PK;
-        expectedCard: Partial<Card>;
-      }>([
-        {
-          columnName: "USER_ID",
-          expectedCard: {
-            dataset_query: {
-              database: SAMPLE_DB_ID,
-              query: {
-                filter: [
-                  "=",
-                  [
-                    "field",
-                    ORDERS.USER_ID,
-                    {
-                      "base-type": "type/Integer",
-                    },
-                  ],
-                  ORDERS_ROW_VALUES.USER_ID,
-                ],
-                "source-table": ORDERS_ID,
-              },
-              type: "query",
-            },
-            display: "table",
-          },
-        },
-        {
-          columnName: "PRODUCT_ID",
-          expectedCard: {
-            dataset_query: {
-              database: SAMPLE_DB_ID,
-              query: {
-                filter: [
-                  "=",
-                  [
-                    "field",
-                    ORDERS.PRODUCT_ID,
-                    {
-                      "base-type": "type/Integer",
-                    },
-                  ],
-                  ORDERS_ROW_VALUES.PRODUCT_ID,
-                ],
-                "source-table": ORDERS_ID,
-              },
-              type: "query",
-            },
-            display: "table",
-          },
-        },
-        {
-          columnName: "CREATED_AT",
-          expectedCard: {
-            dataset_query: {
-              database: SAMPLE_DB_ID,
-              query: {
-                filter: [
-                  "=",
-                  [
-                    "field",
-                    ORDERS.USER_ID,
-                    {
-                      "base-type": "type/Integer",
-                    },
-                  ],
-                  ORDERS_ROW_VALUES.USER_ID,
-                ],
-                "source-table": ORDERS_ID,
-              },
-              type: "query",
-            },
-            display: "table",
-          },
-        },
-      ])(
-        "should apply drill on $columnName cell click",
-        async ({ columnName, expectedCard }) => {
-          const { props } = await setup({
-            question: ORDERS_QUESTION_WITH_MULTIPLE_PK,
-            clicked: {
-              column: ORDERS_COLUMNS_WITH_MULTIPLE_PK[columnName],
-              value: ORDERS_ROW_VALUES_WITH_MULTIPLE_PK[columnName],
-            },
-          });
-
-          const drill = screen.getByText("View details");
-          expect(drill).toBeInTheDocument();
-
-          userEvent.click(drill);
-
-          expect(props.onChangeCardAndRun).toHaveBeenCalledTimes(1);
-          expect(props.onChangeCardAndRun).toHaveBeenLastCalledWith(
-            expect.objectContaining({
-              nextCard: expect.objectContaining(expectedCard),
-            }),
-          );
-        },
-      );
-    });
   });
 });
-
-async function setup({
-  question = Question.create({
-    metadata: SAMPLE_METADATA,
-    dataset_query: ORDERS_DATASET_QUERY,
-  }),
-  clicked,
-  settings = {},
-  dimension: inputDimension,
-  columns = ORDERS_COLUMNS_LIST,
-  rowValues = ORDERS_ROW_VALUES,
-}: Partial<{
-  question: Question;
-  clicked: ClickObject;
-  settings: Record<string, any>;
-  dimension?: Dimension;
-  columns?: DatasetColumn[];
-  rowValues?: Record<string, RowValue>;
-}> = {}) {
-  const mode = checkNotNull(getMode(question));
-
-  const dimension =
-    inputDimension ||
-    (question?.query() as StructuredQuery).dimensionForColumn(
-      checkNotNull(clicked?.column),
-    );
-
-  clicked = {
-    ...clicked,
-    dimension: dimension || undefined,
-  };
-
-  const clickActions = mode.actionsForClick(
-    clicked,
-    settings,
-  ) as RegularClickAction[];
-
-  const dispatch = jest.fn();
-  const onChangeCardAndRun = jest.fn();
-  const onClose = jest.fn();
-  const onUpdateVisualizationSettings = jest.fn();
-
-  // used only in FormatDrill. To be refactored. I think we should pass this widget from the outside, ready to be rendered
-  const series: Series = [
-    createMockSingleSeries(
-      {
-        dataset_query: question.datasetQuery(),
-      },
-      {
-        data: {
-          cols: [...columns],
-          rows: [],
-          requested_timezone: "UTC",
-          results_timezone: "Asia/Nicosia",
-          results_metadata: {
-            columns: [...columns],
-          },
-        },
-      },
-    ),
-  ];
-
-  const view = renderWithProviders(
-    <ClickActionsPopover
-      clickActions={clickActions}
-      clicked={clicked}
-      series={series}
-      dispatch={dispatch}
-      onChangeCardAndRun={onChangeCardAndRun}
-      onUpdateVisualizationSettings={onUpdateVisualizationSettings}
-      onClose={onClose}
-    />,
-  );
-
-  dispatch.mockImplementation(fn => {
-    if (typeof fn === "function") {
-      return fn(dispatch, view.store.getState);
-    }
-  });
-
-  const updatedClicked = {
-    ...clicked,
-    element: view.baseElement,
-  };
-
-  view.rerender(
-    <ClickActionsPopover
-      clickActions={clickActions}
-      clicked={updatedClicked}
-      series={series as unknown as Series}
-      dispatch={dispatch}
-      onChangeCardAndRun={onChangeCardAndRun}
-      onUpdateVisualizationSettings={onUpdateVisualizationSettings}
-      onClose={onClose}
-    />,
-  );
-
-  await waitFor(async () => {
-    expect(
-      (await screen.findAllByTestId("drill-through-section")).length,
-    ).toBeGreaterThan(0);
-  });
-
-  return {
-    props: {
-      clickActions,
-      clicked: updatedClicked,
-      series,
-      dispatch,
-      onChangeCardAndRun,
-      onUpdateVisualizationSettings,
-      onClose,
-    },
-    view,
-  };
-}
 
 function getSummarizedOverTimeResultDatasetQuery(
   aggregatedColumnId: number,
